+++ src/components/forms/styles.ts
@@ -38,7 +38,8 @@ export const checkRadioStyles = baseFormStyles.extend({
 	variants: {
 		state: {
 			default:
-				'border-muted checked:bg-primary indeterminate:bg-primary ' +
+				'border-primary/50 checked:bg-primary checked:border-primary ' +
+				'indeterminate:bg-primary indeterminate:border-primary ' +
 				'focus:ring-primary focus:border-primary',
 			error: 'border-danger focus:ring-danger focus:border-danger',
 		},